+++ devapp-2.16/client/main.jsx
@@ -3,6 +3,7 @@ import { Meteor } from 'meteor/meteor'
 import { render } from 'react-dom'
 import { App } from '../imports/ui/App'
 import './local-collections'
+import './additional-connection'
 
 import '../imports/api/links'
 import '../imports/api/random'