+++ devapp-3.5/client/main.jsx
@@ -3,6 +3,7 @@ import { Meteor } from 'meteor/meteor'
 import { App } from '/imports/ui/App'
 import '/imports/ui/styles.css'
 import './local-collections'
+import './additional-connection'
 
 Meteor.startup(() => {
   const container = document.getElementById('react-target')