+++ src/Browser/MeteorLibrary.ts
@@ -1,10 +1,9 @@
 import { JSONUtils } from '@/Utils/JSONUtils'
 import { mapValues, omit } from '@/Utils/Objects'
 
-export const getSubscriptions = () => {
-  const payload = mapValues(
-    Meteor?.connection?._subscriptions ?? {},
-    (value: any) => omit(value, ['connection', 'readyDeps']),
+export const getSubscriptions = (connection: DDPConnection) => {
+  const payload = mapValues(connection._subscriptions ?? {}, (value: any) =>
+    omit(value, ['connection', 'readyDeps']),
   )
 
   return JSONUtils.stringify(payload)