+++ src/Injectors/MinimongoInjector.ts
@@ -1,4 +1,3 @@
-import { warning } from '@/Log'
 import { Registry, sendMessage } from '@/Browser/Inject'
 import { cleanupDocument } from '@/Utils/Minimongo'
 import {
@@ -7,6 +6,7 @@ import {
   type LocalCollectionLike,
   type MongoCollectionLike,
 } from './CollectionRegistry'
+import { getMeteorConnections } from './MeteorConnections'
 
 const collectionRegistry = createCollectionRegistry()
 
@@ -14,37 +14,6 @@ const getDocs = (collection: LocalCollectionLike) => {
   return collection._docs._map instanceof Map
     ? collection._docs._map?.values() || []
     : Object.values(collection._docs._map || {})
-}
-
-const getCollections = () => {
-  const collections = Meteor.connection._mongo_livedata_collections
-
-  if (!collections) {
-    warning(
-      'Collections not initialized in the client yet. Possibly forgotten to be imported.',
-    )
-    return
-  }
-
-  const registeredCollections = collectionRegistry.list(collections)
-  const data: MinimongoSnapshotPayload = {
-    collections: Object.fromEntries(
-      registeredCollections.map(({ collection, displayName }) => [
-        displayName,
-        [...getDocs(collection)].map(
-          item => cleanupDocument(item) as IDocument,
-        ),
-      ]),
-    ),
-    metadata: Object.fromEntries(
-      registeredCollections.map(({ actualName, displayName }) => [
-        displayName,
-        { actualName },
-      ]),
-    ),
-  }
-
-  sendMessage('minimongo-get-collections', data)
 }
 
 interface MongoNamespace {
@@ -57,7 +26,36 @@ export const MinimongoInjector = () => {
     Mongo as unknown as MongoNamespace,
     collectionRegistry,
   )
-  Registry.register('minimongo-get-collections', () => {
-    getCollections()
+  Registry.register('minimongo-get-collections', message => {
+    const { connectionId } = message.data as ConnectionRequest
+    const descriptor = getMeteorConnections().get(connectionId)
+
+    if (!descriptor) return
+
+    const collections = descriptor.connection._mongo_livedata_collections ?? {}
+    const registeredCollections = collectionRegistry.list(collections, {
+      connection: descriptor.connection,
+      includeUnmanaged: connectionId === 'default',
+    })
+
+    const data: MinimongoSnapshotPayload = {
+      connectionId,
+      collections: Object.fromEntries(
+        registeredCollections.map(({ collection, displayName }) => [
+          displayName,
+          [...getDocs(collection)].map(
+            item => cleanupDocument(item) as IDocument,
+          ),
+        ]),
+      ),
+      metadata: Object.fromEntries(
+        registeredCollections.map(({ actualName, displayName }) => [
+          displayName,
+          { actualName },
+        ]),
+      ),
+    }
+
+    sendMessage('minimongo-get-collections', data)
   })
 }